+++ lib/flutter_map_tile_caching.dart
@@ -2,7 +2,7 @@
 // You should never need to include anything other than this file from this library in your project
 
 export 'src/regions/circle.dart';
-export 'src/regions/downloadableRegion.dart' show RegionType, DownloadProgress;
+export 'src/regions/downloadableRegion.dart';
 export 'src/regions/rectangle.dart';
 
 export 'src/shapeChooser.dart';
@@ -11,3 +11,7 @@ export 'src/preloadSurroundings.dart';
 
 export 'src/main.dart';
 export 'src/storageManager.dart';
+
+export 'src/deprecated/oldCachingManager.dart';
+
+export 'package:network_to_file_image/network_to_file_image.dart';