+++ lib/src/main.dart
@@ -7,13 +7,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter/painting.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:flutter_map/plugin_api.dart';
-import 'package:flutter_map_tile_caching/src/storageManager.dart';
 import 'package:http/http.dart' as http;
 import 'package:latlong2/latlong.dart';
 import 'package:meta/meta.dart';
 import 'package:network_to_file_image/network_to_file_image.dart';
-import 'package:path_provider/path_provider.dart';
 import 'package:permission_handler/permission_handler.dart';
+import 'package:path/path.dart' as p show joinAll;
 
 import 'regions/downloadableRegion.dart';
 
@@ -65,19 +64,34 @@ class CachedNotAvailable implements Exception {
 ///
 /// See online documentation for more information about the caching/downloading behaviour of this library.
 class StorageCachingTileProvider extends TileProvider {
-  /// The maximum number of downloadable tiles per store. Defaults to 20000 (`kMaxTileDownloadCount`).
-  final int maxTileDownloadCount;
+  /// Deprecated. Will be removed in the next release. Migrate to `maxStoreLength`.
+  @Deprecated(
+    'This variable has been deprecated, and will be removed in the next release. Migrate to `maxStoreLength`.',
+  )
+  static final kMaxPreloadTileAreaCount = 20000;
 
-  /// The default maximum number of downloadable tiles per store. Defaults to 20000.
-  static final int kMaxTileDownloadCount = 20000;
+  /// The maximum number of tiles allowed in a cache store ('browsing', see below) before the oldest tile gets deleted.
+  ///
+  /// Only applies to 'browse caching', ie. downloading regions will bypass this limit. Please note that this can be computationally expensive as it potentially involves sorting through this many files to find the oldest file.
+  ///
+  /// Defaults to 20000, set to 0 to disable.
+  final int maxStoreLength;
 
-  /// The duration until a tile expires and needs to be fetched again. Defaults to 31 days.
+  /// The duration until a tile expires and needs to be fetched again when browsing.
+  ///
+  /// Only has an effect when the `CacheBehavior` is set to `cacheFirst`.
+  ///
+  /// Defaults to 31 days, set to 0 to disable.
   final Duration cachedValidDuration;
 
-  /// The name of the cache store to use for this instance. Defaults to the default cache, 'mainCache'.
+  /// The name of the cache store to use for this instance. Defaults to the default store, 'mainCache'.
   final String storeName;
 
-  /// The directory to place cache stores into. Use `await StorageCachingTileProvider.normalDirectory` wherever possible. Required.
+  /// The directory to place cache stores into.
+  ///
+  /// Use `await MapStorageManager.normalDirectory` wherever possible, or `await MapStorageManager.temporaryDirectory` is required (see documentation). If creating a path manually, be sure it's the correct format, use the `path` library if needed.
+  ///
+  /// Required.
   final Directory parentDirectory;
 
   /// The behavior method to get the tile. Defaults to `cacheFirst`.
@@ -91,7 +105,7 @@ class StorageCachingTileProvider extends TileProvider {
   ///
   /// See online documentation for more information about the caching/downloading behaviour of this library.
   StorageCachingTileProvider({
-    this.maxTileDownloadCount = 20000,
+    this.maxStoreLength = 20000,
     this.cachedValidDuration = const Duration(days: 31),
     this.storeName = 'mainCache',
     this.behavior = CacheBehavior.cacheFirst,
@@ -99,22 +113,6 @@ class StorageCachingTileProvider extends TileProvider {
   });
 
   //! GETTER FUNCTIONS !//
-
-  /// Get the application's documents directory
-  ///
-  /// Caching in here will show caches under the App Storage - instead of under App Cache - in Settings, and therefore the OS or other apps cannot clear the cache without telling the user.
-  static Future<Directory> get normalDirectory async {
-    return await getApplicationDocumentsDirectory();
-  }
-
-  /// Get the temporary storage directory
-  ///
-  /// Caching in here will show caches under the App Cache - instead of App Storage - in Settings. Therefore the OS can clear cached tiles at any time without telling the user.
-  ///
-  /// For this reason, it is not recommended to use this store. Use `normalDirectory` by default instead.
-  static Future<Directory> get unstableDirectory async {
-    return await getTemporaryDirectory();
-  }
 
   Future<void> _getAndSaveTile(
     Coords<num> coord,
@@ -126,27 +124,66 @@ class StorageCachingTileProvider extends TileProvider {
         Coords(coord.x.toDouble(), coord.y.toDouble())..z = coord.z.toDouble();
     final String url = getTileUrl(coordDouble, options);
 
-    await TileStorageManager(parentDirectory, storeName).newTile(
-      url: url,
-      client: client,
-      errorHandler: (e) => errorHandler(url, e),
-    );
+    try {
+      Directory(p.joinAll([parentDirectory.path, storeName]))
+          .createSync(recursive: true);
+      File(
+        p.joinAll([
+          parentDirectory.path,
+          storeName,
+          url
+              .replaceAll('https://', '')
+              .replaceAll('http://', '')
+              .replaceAll("/", "")
+              .replaceAll(".", "")
+        ]),
+      ).writeAsBytesSync((await client.get(Uri.parse(url))).bodyBytes);
+    } catch (e) {
+      errorHandler(url, e);
+    }
   }
 
   /// Get a browsed tile as an image, paint it on the map and save it's bytes to cache for later
   @override
   ImageProvider getImage(Coords<num> coords, TileLayerOptions options) {
-    final File fileReal = File(
-      parentDirectory.path +
-          '/$storeName/tiles/' +
-          getTileUrl(coords, options)
-              .replaceAll('https://', '')
-              .replaceAll('http://', '')
-              .replaceAll("/", ""),
-    );
+    final File fileReal = File(p.joinAll([
+      parentDirectory.path,
+      storeName,
+      getTileUrl(coords, options)
+          .replaceAll('https://', '')
+          .replaceAll('http://', '')
+          .replaceAll("/", "")
+          .replaceAll(".", "")
+    ]));
 
     NetworkToFileImage ntfi(bool url, bool file) {
       try {
+        List<FileSystemEntity>? fileList;
+        if (maxStoreLength != 0 &&
+            file &&
+            url &&
+            Directory(p.joinAll([parentDirectory.path, storeName]))
+                .existsSync())
+          fileList = Directory(p.joinAll([parentDirectory.path, storeName]))
+              .listSync();
+
+        if (fileList != null &&
+            fileList.isNotEmpty &&
+            fileList.length + 1 > maxStoreLength) {
+          /*FileSystemEntity? currentOldestFile;
+          DateTime currentOldestTime = DateTime(9999);
+          fileList.forEach((element) {
+            DateTime mod = element.statSync().modified;
+            if (mod.isBefore(currentOldestTime)) {
+              currentOldestFile = element;
+              currentOldestTime = mod;
+            }
+          });
+          currentOldestFile!.deleteSync();*/
+          fileList.sort(
+              (a, b) => b.statSync().modified.compareTo(a.statSync().modified));
+          fileList[0].deleteSync();
+        }
         return NetworkToFileImage(
           url: url ? getTileUrl(coords, options) : null,
           file: file ? fileReal : null,
@@ -156,28 +193,31 @@ class StorageCachingTileProvider extends TileProvider {
       }
     }
 
-    if (behavior == CacheBehavior.onlineOnly) {
-      return ntfi(true, false);
-    } else if (behavior == CacheBehavior.cacheOnly) {
-      if (!Directory(parentDirectory.path + '/$storeName/tiles/').existsSync())
-        throw CachedNotAvailable();
-      return ntfi(false, true);
-    } else {
-      Directory(parentDirectory.path + '/$storeName/tiles/')
-          .createSync(recursive: true);
+    Directory(p.joinAll([parentDirectory.path, storeName]))
+        .createSync(recursive: true);
 
+    if (behavior == CacheBehavior.onlineOnly)
+      return ntfi(true, false);
+    else if (behavior == CacheBehavior.cacheOnly)
+      return ntfi(false, true);
+    else {
       if (behavior == CacheBehavior.cacheRenew ||
           (fileReal.existsSync() &&
               fileReal
                   .lastModifiedSync()
                   .add(cachedValidDuration)
                   .isBefore(DateTime.now())))
-        TileStorageManager(parentDirectory, storeName).deleteTile(
-          fileName: getTileUrl(coords, options)
-              .replaceAll('https://', '')
-              .replaceAll('http://', '')
-              .replaceAll("/", ""),
-        );
+        File(
+          p.joinAll([
+            parentDirectory.path,
+            storeName,
+            getTileUrl(coords, options)
+                .replaceAll('https://', '')
+                .replaceAll('http://', '')
+                .replaceAll("/", "")
+                .replaceAll(".", "")
+          ]),
+        ).deleteSync();
       return ntfi(true, true);
     }
   }
@@ -274,21 +314,24 @@ class StorageCachingTileProvider extends TileProvider {
 
   /// Download a specified `DownloadableRegion` in the background, and show a notification progress bar (by default)
   ///
-  /// Only available on Android devices, due to limitations with other operating systems.
+  /// Only available on Android devices, due to limitations with other operating systems. This may be slower and/or less stable than conducting the download in the foreground.
   ///
   /// To check the number of tiles that need to be downloaded before using this function, use `checkRegion()`.
   ///
-  /// You may want to call `requestIgnoreBatteryOptimizations()` before hand, depending on how/where/why this background download will be used. See documentation on that function for more information.
+  /// You may want to call `requestIgnoreBatteryOptimizations()` before hand, depending on how/where/why this background download will be used. See documentation on that method for more information.
   ///
-  /// Optionally specify a `callback` that gets fired every time another tile is downloaded/failed, takes one `DownloadProgress` argument, and returns a boolean.
+  /// Optionally specify `showNotification` as `false` to disable the built-in notification system.
   ///
-  /// Download can be cancelled by returning `true` from `callback` function or by fully closing the app.
+  /// Optionally specify a `callback` that gets fired every time another tile is downloaded/failed, takes one `DownloadProgress` argument, and returns a boolean. Download can be cancelled by returning `true` from `callback` function or by fully closing the app.
+  ///
+  /// If the download doesn't seem to start on a testing device, try changing `useAltMethod` to `true`. This will switch to an older Android API, so should only be used if it is the most stable on a device. You may be able to programatically detect if the download hasn't started by using the callback, therefore allowing you to call this method again with `useAltMethod`, but this isn't guranteed.
   ///
   /// Returns nothing.
   void downloadRegionBackground(
     DownloadableRegion region, {
     bool showNotification = true,
     bool Function(DownloadProgress)? callback,
+    bool useAltMethod = false,
   }) async {
     if (io.Platform.isAndroid) {
       FlutterLocalNotificationsPlugin? flutterLocalNotificationsPlugin;
@@ -376,7 +419,7 @@ class StorageCachingTileProvider extends TileProvider {
         TaskConfig(
           taskId: 'backgroundTileDownload',
           delay: 1,
-          forceAlarmManager: true,
+          forceAlarmManager: useAltMethod,
         ),
       );
     } else
@@ -401,17 +444,6 @@ class StorageCachingTileProvider extends TileProvider {
       outlineTileNums[zoomLvl] = {};
 
       for (LatLng node in circleOutline) {
-        /*
-        The below code has been retained on purpose.
-        DO NOT remove it.
-        
-        final double n = math.pow(2, zoomLvl).toDouble();
-        final int x = ((node.longitude + 180.0) / 360.0 * n).toInt();
-        final int y =
-            ((1.0 - _asinh(math.tan(node.latitudeInRad)) / math.pi) / 2.0 * n)
-                .toInt();
-        */
-
         final CustomPoint<num> tile = crs
             .latLngToPoint(node, zoomLvl.toDouble())
             .unscaleBy(tileSize)
@@ -464,8 +496,6 @@ class StorageCachingTileProvider extends TileProvider {
       crs: crs,
       tileSize: tileSize,
     );
-    assert(tiles.length <= maxTileDownloadCount,
-        '${tiles.length} exceeds maximum number of pre-cacheable tiles');
 
     final List<String> erroredUrls = [];
     final http.Client client = http.Client();
@@ -532,8 +562,6 @@ class StorageCachingTileProvider extends TileProvider {
       crs: crs,
       tileSize: tileSize,
     );
-    assert(tiles.length <= maxTileDownloadCount,
-        '${tiles.length} exceeds maximum number of pre-cacheable tiles');
 
     final List<String> erroredUrls = [];
     final http.Client client = http.Client();