+++ lib/src/storageManager.dart
@@ -1,80 +1,115 @@
 import 'dart:io';
 
-import 'package:http/http.dart' as http;
+import 'package:path_provider/path_provider.dart';
+import 'package:path/path.dart' as p show joinAll;
+import 'exts.dart';
 
 /// Handles caching for tiles
 ///
-/// Used internally for downloading regions, another library is depended on for browse caching
-class TileStorageManager {
-  /// The directory to place cache stores into. Use the same directory used in `StorageCachingTileProvider` wherever possible. Required.
+/// Used internally for downloading regions, another library is depended on for 'browse caching'.
+class MapCachingManager {
+  /// The directory to place cache stores into
+  ///
+  /// Use the same directory used in `StorageCachingTileProvider` (`await MapCachingManager.normalDirectory` recommended). Required.
   final Directory parentDirectory;
 
   /// The name of a store. Defaults to 'mainCache'.
   final String storeName;
 
+  /// The correctly joined `parentDirectory` and `storeName` at time of initialization.
+  final String _joinedBasePath;
+
   /// Create an instance to handle caching for tiles
   ///
-  /// Used internally for downloading regions, another library is depended on for browse caching
-  TileStorageManager(this.parentDirectory, [this.storeName = 'mainCache']);
+  /// Used internally for downloading regions, another library is depended on for 'browse caching'.
+  MapCachingManager(this.parentDirectory, [this.storeName = 'mainCache'])
+      : _joinedBasePath = p.joinAll([parentDirectory.path, storeName]);
 
-  /// Cache a new tile given at least a URL, and optionally an HTTP client and an error handler function
-  Future<void> newTile({
-    required String url,
-    http.Client? client,
-    void Function(dynamic)? errorHandler,
-  }) async {
-    try {
-      Directory(parentDirectory.path + '/$storeName/tiles/')
-          .createSync(recursive: true);
-      File(
-        parentDirectory.path +
-            '/$storeName/tiles/' +
-            url
-                .replaceAll('https://', '')
-                .replaceAll('http://', '')
-                .replaceAll("/", ""),
-      ).writeAsBytesSync(
-          (await (client ?? http.Client()).get(Uri.parse(url))).bodyBytes);
-    } catch (e) {
-      if (errorHandler != null) errorHandler(e);
-    }
-  }
-
-  /// Delete a cached tile given a URL or a file name
-  void deleteTile({
-    String? url,
-    String? fileName,
-  }) {
-    assert(
-      (!(url == null && fileName == null)) ||
-          (!(url != null && fileName != null)),
-      'Provide one of either `url` or `fileName` to delete a tile',
-    );
-    final file = (url != null
-        ? File(
-            parentDirectory.path +
-                '/$storeName/tiles/' +
-                url
-                    .replaceAll('https://', '')
-                    .replaceAll('http://', '')
-                    .replaceAll("/", ""),
-          )
-        : File(
-            parentDirectory.path + '/$storeName/tiles/' + fileName!,
-          ));
-    if (file.existsSync()) file.deleteSync();
-  }
-
-  /// Delete a whole cache store
+  /// Delete a cache store
   void deleteStore() {
-    if (Directory(parentDirectory.path + '/$storeName/').existsSync())
-      Directory(parentDirectory.path + '/$storeName/')
-          .deleteSync(recursive: true);
+    if (Directory(_joinedBasePath).existsSync())
+      Directory(_joinedBasePath).deleteSync(recursive: true);
   }
 
   /// Delete all cache stores
-  void clearCache() {
+  void deleteAllStores() {
     if (Directory(parentDirectory.path).existsSync())
       Directory(parentDirectory.path).deleteSync(recursive: true);
+  }
+
+  /// Retrieve a list of all names of existing cache stores
+  ///
+  /// Returns null if the cache does not exist.
+  List<String>? get allStoresNames {
+    if (!Directory(parentDirectory.path).existsSync()) return null;
+    List<String> returnable = [];
+    for (FileSystemEntity dir in Directory(parentDirectory.path)
+        .listSync(followLinks: false, recursive: false)) {
+      returnable.add(dir.path.split('/')[dir.path.split('/').length - 2]);
+    }
+    return returnable;
+  }
+
+  /// Retrieve the size (in bytes) of a cache store
+  ///
+  /// Use `.bytesToMegabytes` on the output to get the real number of megabytes
+  ///
+  /// Returns null if the store does not exist.
+  int? get storeSize {
+    print(_joinedBasePath);
+    if (!Directory(_joinedBasePath).existsSync()) return null;
+    return Directory(_joinedBasePath).statSync().size;
+  }
+
+  /// Retrieve the size (in bytes) of all cache stores
+  ///
+  /// Use `.bytesToMegabytes` on the output to get the real number of megabytes.
+  ///
+  /// Returns null if the cache does not exist.
+  int? get allStoresSize {
+    if (!Directory(parentDirectory.path).existsSync()) return null;
+    return Directory(parentDirectory.path).statSync().size;
+  }
+
+  /// Retrieve the number of stored tiles in a cache store
+  ///
+  /// Returns null if the store does not exist.
+  int? get storeLength {
+    if (!Directory(_joinedBasePath).existsSync()) return null;
+    return Directory(_joinedBasePath).listSync().length;
+  }
+
+  /// Retrieve the (approximate) number of stored tiles in all cache stores
+  ///
+  /// Returns null if the cache does not exist.
+  int? get allStoresLength {
+    if (!Directory(parentDirectory.path).existsSync()) return null;
+    return (Directory(parentDirectory.path).listSync(recursive: true).length) -
+        ((allStoresNames?.length ?? 0) * 2);
+  }
+
+  /// Get the application's documents directory
+  ///
+  /// Caching in here will show caches under the App Storage - instead of under App Cache - in Settings, and therefore the OS or other apps cannot clear the cache without telling the user.
+  static Future<Directory> get normalDirectory async {
+    return Directory(p.joinAll(
+        [(await getApplicationDocumentsDirectory()).path, 'mapCache']));
+  }
+
+  /// Get the temporary storage directory
+  ///
+  /// Caching in here will show caches under the App Cache - instead of App Storage - in Settings. Therefore the OS can clear cached tiles at any time without telling the user.
+  ///
+  /// For this reason, it is not recommended to use this store. Use `normalDirectory` by default instead.
+  static Future<Directory> get temporaryDirectory async {
+    return Directory(
+        p.joinAll([(await getTemporaryDirectory()).path, 'mapCache']));
+  }
+
+  /// Returns the current working store's path, store's size, and all stores size, as a human-friendly 3-line string
+  @override
+  String toString() {
+    final String path = p.joinAll([parentDirectory.path, storeName]);
+    return '$path\nStore size: ${(storeSize ?? 0).bytesToMegabytes}MB\nAll stores size: ${(allStoresSize ?? 0).bytesToMegabytes}MB';
   }
 }